Refactor the roundup report into its own class
